+++ fuel-tx/src/tests/bytes.rs
@@ -446,7 +446,7 @@ fn transaction_serde_serialization_deserialization() {
         ),
         Transaction::upgrade(
             UpgradePurpose::StateTransition {
-                bytecode_hash: [0xfa; 32].into(),
+                root: [0xfa; 32].into(),
             },
             Policies::new()
                 .with_tip(Word::MAX >> 1)
@@ -459,7 +459,7 @@ fn transaction_serde_serialization_deserialization() {
         ),
         Transaction::upgrade(
             UpgradePurpose::StateTransition {
-                bytecode_hash: [0xfa; 32].into(),
+                root: [0xfa; 32].into(),
             },
             Policies::new()
                 .with_tip(Word::MAX >> 1)
@@ -472,7 +472,7 @@ fn transaction_serde_serialization_deserialization() {
         ),
         Transaction::upgrade(
             UpgradePurpose::StateTransition {
-                bytecode_hash: [0xfa; 32].into(),
+                root: [0xfa; 32].into(),
             },
             Policies::new()
                 .with_tip(Word::MAX >> 1)
@@ -485,7 +485,77 @@ fn transaction_serde_serialization_deserialization() {
         ),
         Transaction::upgrade(
             UpgradePurpose::StateTransition {
-                bytecode_hash: [0xfa; 32].into(),
+                root: [0xfa; 32].into(),
+            },
+            Policies::new()
+                .with_tip(Word::MAX >> 1)
+                .with_maturity((u32::MAX >> 3).into())
+                .with_witness_limit(Word::MAX >> 4)
+                .with_max_fee(Word::MAX >> 5),
+            vec![],
+            vec![],
+            vec![],
+        ),
+    ]);
+    assert_encoding_correct(&[
+        Transaction::upload(
+            UploadBody {
+                root: [6; 32].into(),
+                witness_index: 0,
+                subsection_index: 0x1234,
+                subsections_number: 0x4321,
+                proof_set: vec![[1; 32].into(), [2; 32].into(), [3; 32].into()],
+            },
+            Policies::new()
+                .with_tip(Word::MAX >> 1)
+                .with_maturity((u32::MAX >> 3).into())
+                .with_witness_limit(Word::MAX >> 4)
+                .with_max_fee(Word::MAX >> 5),
+            vec![i.clone()],
+            vec![o],
+            vec![w.clone()],
+        ),
+        Transaction::upload(
+            UploadBody {
+                root: [6; 32].into(),
+                witness_index: 0,
+                subsection_index: 0x1234,
+                subsections_number: 0x4321,
+                proof_set: vec![[1; 32].into(), [2; 32].into(), [3; 32].into()],
+            },
+            Policies::new()
+                .with_tip(Word::MAX >> 1)
+                .with_maturity((u32::MAX >> 3).into())
+                .with_witness_limit(Word::MAX >> 4)
+                .with_max_fee(Word::MAX >> 5),
+            vec![],
+            vec![o],
+            vec![w.clone()],
+        ),
+        Transaction::upload(
+            UploadBody {
+                root: [6; 32].into(),
+                witness_index: 0,
+                subsection_index: 0x1234,
+                subsections_number: 0x4321,
+                proof_set: vec![[1; 32].into(), [2; 32].into(), [3; 32].into()],
+            },
+            Policies::new()
+                .with_tip(Word::MAX >> 1)
+                .with_maturity((u32::MAX >> 3).into())
+                .with_witness_limit(Word::MAX >> 4)
+                .with_max_fee(Word::MAX >> 5),
+            vec![],
+            vec![],
+            vec![w.clone()],
+        ),
+        Transaction::upload(
+            UploadBody {
+                root: [6; 32].into(),
+                witness_index: 0,
+                subsection_index: 0x1234,
+                subsections_number: 0x4321,
+                proof_set: vec![[1; 32].into(), [2; 32].into(), [3; 32].into()],
             },
             Policies::new()
                 .with_tip(Word::MAX >> 1)