+++ fuel-vm/src/storage/predicate.rs
@@ -230,14 +230,7 @@ impl InterpreterStorage for PredicateStorage {
         &mut self,
         _version: u32,
         _consensus_parameters: &ConsensusParameters,
-    ) -> Result<Option<Cow<'_, ConsensusParameters>>, Self::DataError> {
-        Err(StorageUnavailable)
-    }
-
-    fn contains_state_transition_bytecode_hash(
-        &self,
-        _hash: &Bytes32,
-    ) -> Result<bool, Self::DataError> {
+    ) -> Result<Option<ConsensusParameters>, Self::DataError> {
         Err(StorageUnavailable)
     }
 
@@ -245,7 +238,7 @@ impl InterpreterStorage for PredicateStorage {
         &mut self,
         _version: u32,
         _hash: &Bytes32,
-    ) -> Result<Option<Cow<'_, Bytes32>>, Self::DataError> {
+    ) -> Result<Option<Bytes32>, Self::DataError> {
         Err(StorageUnavailable)
     }