+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Breaking
 
+- [#715](https://github.com/FuelLabs/fuel-vm/pull/715): The `Interpreter` supports the processing of the `Upload` transaction. The change affects `InterpreterStorage`, adding `StorageMutate<UploadedBytes>` constrain.
 - [#714](https://github.com/FuelLabs/fuel-vm/pull/714): The change adds a new `Upload` transaction that allows uploading huge byte code on chain subsection by subsection. This transaction is chargeable and is twice as expensive as the `Create` transaction. Anyone can submit this transaction.
 - [#712](https://github.com/FuelLabs/fuel-vm/pull/712): The `Interpreter` supports the processing of the `Upgrade` transaction. The change affects `InterpreterStorage`, adding 5 new methods that must be implemented.
 - [#707](https://github.com/FuelLabs/fuel-vm/pull/707): The change adds a new `Upgrade` transaction that allows upgrading either consensus parameters or state transition function used by the network to produce future blocks.
@@ -33,10 +34,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
         },
         /// The upgrade is performed to change the state transition function.
         StateTransition {
-            /// The hash of the new bytecode of the state transition function.
+            /// The Merkle root of the new bytecode of the state transition function.
             /// The bytecode must be present on the blockchain(should be known by the
             /// network) at the moment of inclusion of this transaction.
-            bytecode_hash: Bytes32,
+            root: Bytes32,
         },
     }
     ```